backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pandas
google-generativeai
